hwjs_3

+++ js/task-3.js
@@ -1,7 +1,7 @@
 function getElementWidth(content, padding, border) {
-    const contentWidth = parseInt(content);
-    const paddingWidth = parseInt(padding);
-    const borderThickness = parseInt(border);
+    const contentWidth = parseFloat(content);
+    const paddingWidth = parseFloat(padding);
+    const borderThickness = parseFloat(border);
     
     return contentWidth + 2 * paddingWidth + 2 * borderThickness;
 }
